fix for issue 2, swap failure
* On Fedora 19/20 the os requires that the swap file have a permission
setting of at least 600. The script was using the default umask
which was setting the permissions to 644. The swap function now sets
the permissions of the swap file to 600.

* In addition to the permission issue i discovered that in Fedora
19/20 the swap file can not exist in "/tmp", to correct this I moved
the swap file to "/".

* Lastly, because i was making changes to the script I decied to
reduced the swap file size from 2048 to 1024. I made this last change
because 2048 is excessive for a swap file that may never even be used.